+++ src/main/java/com/example/bankingSystem/model/TransferMoneyDTO.java
@@ -4,23 +4,23 @@ package com.example.bankingSystem.model;
  * Created by mandeep.singh on 12:51 PM 23/03/21 Tuesday
  */
 public class TransferMoneyDTO {
-    long srcAccount;
-    long destAccount;
+    int srcAccount;
+    int destAccount;
     double amount;
 
-    public long getSrcAccount() {
+    public int getSrcAccount() {
         return srcAccount;
     }
 
-    public void setSrcAccount(long srcAccount) {
+    public void setSrcAccount(int srcAccount) {
         this.srcAccount = srcAccount;
     }
 
-    public long getDestAccount() {
+    public int getDestAccount() {
         return destAccount;
     }
 
-    public void setDestAccount(long destAccount) {
+    public void setDestAccount(int destAccount) {
         this.destAccount = destAccount;
     }
